+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+### Fixed
+- fix clipboard not working in webxdc apps
+
 <a id="1_40_4"></a>
 
 ## [1.40.4] - 2023-09-14